+++ docs/source/whatsnew.rst
@@ -4,6 +4,7 @@ What's New
 ==========
 v2.3.4
 ------
+- Many changes to XSIEvents and the classes associated with it. XSIEvents should now be much more stable.
 - Added debug logging to :py:class:`XSIEventsChannel` to help diagnose heartbeat problems
 - :py:meth:`
 - BUG FIX: :py:meth:`Org.get_locations()` no longer creates a duplicate of all locations